Made creating files and also saving for library. There are main problems.
FOR NOW APP CREATE FILES, BUT YOU HAVE TO MANUALLY CREATE LIBRARY AND ADD FILES. Also probelms with Translator.vi It look like you have to copy it from template and then change it for your needs. But when you create in from skretch it's not recognized and compiler show s you error Translator.vi not found.

So now it's only creating VI.

Need to do:
  - create Translator.vi with universal transcripting functions from funcPrototypes.txt